Move post install to its own module
