+++ setup.py
@@ -28,19 +28,16 @@ The t-matrix and mie scattering codes require working Fortran 90 and
 C compilers, as well as f2py and cython. On Ubuntu, you will need the
 "gfortran" and "python-dev" packages installed.
 '''
-import glob
 import os
-import shutil
-import site
-import sys
 from os.path import join
+import sys
 
 import nose
 from numpy.distutils.core import setup, Extension
 from numpy.distutils.misc_util import Configuration
 import setuptools
-from setuptools.command.develop import develop
-from setuptools.command.install import install
+
+from post_install import PostDevelopConfig, PostInstallConfig
 
 try:
     from holopy import __version__
@@ -51,22 +48,6 @@ HOLOPY_NOSE_PLUGIN_LOCATION = ('holopycatchwarnings = '
                                'holopy.core.tests.common:HoloPyCatchWarnings')
 
 hp_root = os.path.dirname(os.path.realpath(sys.argv[0]))
-
-
-class PostDevelopConfig(develop):
-    """Post-installation for development mode."""
-    def run(self):
-        develop.run(self)
-        if os.name == 'nt':
-            _move_msvc_libs('develop')
-
-
-class PostInstallConfig(install):
-    """Post-installation for installation mode."""
-    def run(self):
-        install.run(self)
-        if os.name == 'nt':
-            _move_msvc_libs('install')
 
 
 def configuration(parent_package='', top_path=''):
@@ -92,39 +73,6 @@ def configuration(parent_package='', top_path=''):
     config.add_data_files(join(hp_root, 'holopy', 'propagation', 'tests',
                                'gold', '*.yaml'))
     return config
-
-
-def _move_msvc_libs(mode='install'):
-    """ These dlls need to be moved if the fortran is complied in an
-    environment with MSVC 2015 as the C compiler.
-    """
-    package_dir = _get_holopy_install_dir(mode)
-    lib_dir = join(package_dir, '.libs')
-
-    tmatrix_libs = glob.glob(join(lib_dir, 'libS.*.dll'))
-    mie_libs = glob.glob(join(lib_dir, 'libscsm*.dll'))
-    mie_libs += glob.glob(join(lib_dir, 'libmieang*.dll'))
-    mie_libs += glob.glob(join(lib_dir, 'libuts*.dll'))
-
-    tmatrix_f_dir = join(package_dir, 'scattering', 'theory', 'tmatrix_f')
-    mie_f_dir = join(package_dir, 'scattering', 'theory', 'mie_f')
-
-    for dll in tmatrix_libs:
-        shutil.move(dll, tmatrix_f_dir)
-    for dll in mie_libs:
-        shutil.move(dll, mie_f_dir)
-    shutil.rmtree(lib_dir)
-
-
-def _get_holopy_install_dir(mode):
-    if mode == 'install':
-        hp_dir = [path for path in site.getsitepackages()
-                  if 'site-packages' in path]
-        hp_dir = os.path.join(dir[0], 'holopy')
-    if mode == 'develop':
-        hp_dir = os.path.dirname(os.path.realpath(__file__))
-        hp_dir = os.path.join(dir, 'holopy')
-    return hp_dir
 
 
 if __name__ == "__main__":
